Revert "fix(solvers): cast CVXPy parameters to f64 at the JAX boundary"
This reverts commit 33c9f5771e1cc54fe9d6bbb04226ed466e5ffed7.

+++ openscvx/solvers/cvxpy_ptr_solver.py
@@ -659,11 +659,11 @@ class CVXPyPTRSolver(PTRSolver):
         if "x_prop" in self._problem.param_dict:
             self._set_param("x_prop", x_prop)
         elif self._ocp_vars.x_prop is not None:
-            self._ocp_vars.x_prop.value = np.asarray(x_prop, dtype=np.float64)
+            self._ocp_vars.x_prop.value = np.asarray(x_prop)
         if x_prop_plus is not None and self._ocp_vars.x_prop_plus is not None:
-            self._ocp_vars.x_prop_plus.value = np.asarray(x_prop_plus, dtype=np.float64)
+            self._ocp_vars.x_prop_plus.value = np.asarray(x_prop_plus)
         if E_d is not None and self._ocp_vars.E_d is not None:
-            self._ocp_vars.E_d.value = np.asarray(E_d, dtype=np.float64)
+            self._ocp_vars.E_d.value = np.asarray(E_d)
 
     def update_constraint_linearizations(
         self,
@@ -780,8 +780,7 @@ class CVXPyPTRSolver(PTRSolver):
         """
         try:
             param = self._problem.param_dict[name]
-            # Conic backends require f64; JAX inputs may arrive as f32.
-            value_arr = np.asarray(value, dtype=np.float64)
+            value_arr = np.asarray(value)
 
             # Ensure the value shape matches the parameter shape exactly
             # This is critical for Python 3.11+ where NumPy/CVXPy are stricter about shapes
